Change available field from int to float in LowStockItem schema to properly handle fractional quantities in inventory calculations.

+++ backend/app/schemas/alerts.py
@@ -74,7 +74,7 @@ class LowStockItem(BaseModel):
     """Internal schema for low stock detection."""
     sku_code: str
     sku_name: str
-    available: int
+    available: float
     reorder_point: int
     
     @property
